requirements.txt
Django==1.11.3
django-bootstrap3==9.0.0
django-filter==1.0.4
django-forms-bootstrap==3.1.0
djangorestframework==3.6.3
djutils==0.3.2
Markdown==2.6.8
markovify==0.6.0
nltk==3.2.4
pytz==2017.2
six==1.10.0
Unidecode==0.4.21